+++ packages/remax-cli/src/__tests__/integration/fixtures/typescript/expected/wechat/pages/index.js
@@ -45,7 +45,7 @@ var _page = function _page() {
   var _, _obj$a;
 
   var props = {};
-  var TextElement = react__WEBPACK_IMPORTED_MODULE_1__["cloneElement"]( /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__["createElement"](_components_Text__WEBPACK_IMPORTED_MODULE_3__["default"], null));
+  var TextElement = /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__["cloneElement"]( /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__["createElement"](_components_Text__WEBPACK_IMPORTED_MODULE_3__["default"], null));
 
   var handleClick = function handleClick() {
     return void 0;