+++ packages/remax-cli/src/__tests__/integration/fixtures/wechat/expected/packageA/pages/index.js
@@ -53,7 +53,7 @@ function _extends() { _extends = Object.assign || function (target) { for (var i
 
 var pageA = function pageA() {
   var props = {};
-  var TextElement = react__WEBPACK_IMPORTED_MODULE_1__["cloneElement"]( /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__["createElement"](remax_wechat__WEBPACK_IMPORTED_MODULE_2__["Text"], null));
+  var TextElement = /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__["cloneElement"]( /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__["createElement"](remax_wechat__WEBPACK_IMPORTED_MODULE_2__["Text"], null));
 
   function handleClick() {}